Use Math.toIntExact
Throws exception if we ever go over int max value.

+++ api-src/org/labkey/api/snd/SNDSequencer.java
@@ -37,11 +37,11 @@ public enum SNDSequencer
         minId = id;
     }
 
-    private Integer generateId(Container c)
+    private int generateId(Container c)
     {
         DbSequence sequence = DbSequenceManager.get(c, sequenceName);
         sequence.ensureMinimum(minId);
-        return sequence.next();
+        return Math.toIntExact(sequence.next());
     }
 
     public Integer ensureId(Container container, Integer id)
